refactor: update main README and improve actor imports
