+++ deno_dist/actor/immer/slice-immer.ts
@@ -1,6 +1,6 @@
 import { createDraft, finishDraft } from 'npm:immer@9.0.16';
 
-import { Action, SliceConfig } from '../main/index.ts';
+import { Action, SliceConfig } from '../main/internal.ts';
 import type { Tail } from '../main/utils.ts';
 
 export namespace SliceImmer {
@@ -38,7 +38,7 @@ export namespace SliceImmer {
     applyIncluder: (state, action, includeHandler) => {
       const draft = createDraft(state as any) as typeof state;
       includeHandler(draft, action);
-      return finishDraft(draft) as typeof state;
+      return finishDraft(draft) as any;
     },
     applyHandlerResult: (state, action, result) => result,
     createAction: (sliceName, actionName) =>
