+++ deno_dist/actor/patch/slice-patch.ts
@@ -1,6 +1,6 @@
-import { Patch, patch } from '../../deep/mod.ts';
+import { type Patch, patch } from '../../deep/mod.ts';
 
-import { Action, SliceConfig } from '../main/index.ts';
+import { Action, SliceConfig } from '../main/internal.ts';
 
 export namespace SlicePatch {
   export interface Config extends SliceConfig {
